allowed to have use_ws=True with all other system() classes
