+++ tests/reference/berry_GaAs_W90-conductivity_ohmic_iter-0000.dat
@@ -1,0 +1,12 @@
+#    Efermi                  xx                 xy                 xz                 yx                 yy                 yz                 zx                 zy                 zz                 xx                 xy                 xz                 yx                 yy                 yz                 zx                 zy                 zz        
+   7.000000e+00          -6.381487e+08       4.037086e+03       1.114811e+03       4.037086e+03      -6.381229e+08      -3.188606e+04       1.114811e+03      -3.188606e+04      -2.014106e+09      -7.629443e+08       2.713535e+03       2.995497e+03       2.713535e+03      -7.629189e+08      -3.006658e+04       2.995497e+03      -3.006658e+04      -2.093349e+09
+   7.200000e+00          -2.240992e+09      -1.369302e+04       2.651144e+04      -1.369302e+04      -2.240971e+09      -7.509829e+03       2.651144e+04      -7.509829e+03      -3.041286e+09      -2.278981e+09      -1.162603e+04       2.288753e+04      -1.162603e+04      -2.278958e+09      -1.036024e+04       2.288753e+04      -1.036024e+04      -3.033614e+09
+   7.400000e+00          -4.348453e+09      -1.731252e+03      -2.736347e+01      -1.731252e+03      -4.348417e+09      -2.407753e+04      -2.736347e+01      -2.407753e+04      -3.937775e+09      -4.199044e+09      -2.515740e+03       1.792168e+03      -2.515740e+03      -4.199009e+09      -2.295872e+04       1.792168e+03      -2.295872e+04      -3.874293e+09
+   7.600000e+00          -4.348453e+09      -1.731252e+03      -2.736347e+01      -1.731252e+03      -4.348417e+09      -2.407753e+04      -2.736347e+01      -2.407753e+04      -3.937775e+09      -4.409689e+09      -2.886694e+02      -4.980532e+02      -2.886694e+02      -4.409653e+09      -2.457422e+04      -4.980532e+02      -2.457422e+04      -3.991617e+09
+   7.800000e+00          -5.263101e+09       1.901306e+04      -7.277940e+03       1.901306e+04      -5.263068e+09      -3.148512e+04      -7.277940e+03      -3.148512e+04      -4.721605e+09      -5.197678e+09       1.753246e+04      -6.759779e+03       1.753246e+04      -5.197645e+09      -3.095558e+04      -6.759779e+03      -3.095558e+04      -4.665201e+09
+   8.000000e+00          -5.263101e+09       1.901306e+04      -7.277940e+03       1.901306e+04      -5.263068e+09      -3.148512e+04      -7.277940e+03      -3.148512e+04      -4.721605e+09      -5.213544e+09       1.803926e+04      -6.906478e+03       1.803926e+04      -5.213511e+09      -3.109815e+04      -6.906478e+03      -3.109815e+04      -4.663247e+09
+   8.200000e+00          -4.578698e+09       5.623631e+03      -2.156326e+03       5.623631e+03      -4.578668e+09      -2.614659e+04      -2.156326e+03      -2.614659e+04      -3.909422e+09      -4.624508e+09       6.560063e+03      -2.514907e+03       6.560063e+03      -4.624477e+09      -2.653003e+04      -2.514907e+03      -2.653003e+04      -3.964999e+09
+   8.400000e+00          -4.578698e+09       5.623631e+03      -2.156326e+03       5.623631e+03      -4.578668e+09      -2.614659e+04      -2.156326e+03      -2.614659e+04      -3.909422e+09      -4.462455e+09       5.328369e+03      -2.074204e+03       5.328369e+03      -4.462424e+09      -2.651701e+04      -2.074204e+03      -2.651701e+04      -3.832708e+09
+   8.600000e+00          -2.994040e+09       1.448548e+03      -8.154391e+02       1.448548e+03      -2.994000e+09      -3.120142e+04      -8.154391e+02      -3.120142e+04      -2.839038e+09      -2.837637e+09       7.340881e+02      -1.131957e+03       7.340881e+02      -2.837598e+09      -3.115735e+04      -1.131957e+03      -3.115735e+04      -2.793151e+09
+   8.800000e+00           8.073933e+08      -1.281604e+04      -3.886131e+03      -1.281604e+04       8.074250e+08      -3.555897e+04      -3.886131e+03      -3.555897e+04      -1.125024e+09       5.521426e+08      -1.199272e+04      -3.943987e+03      -1.199272e+04       5.521749e+08      -3.547553e+04      -3.943987e+03      -3.547553e+04      -1.221174e+09
+   9.000000e+00           1.027267e+09      -1.555145e+04      -7.832838e+03      -1.555145e+04       1.027299e+09      -3.891219e+04      -7.832838e+03      -3.891219e+04      -7.437104e+08       1.004482e+09      -1.532127e+04      -7.528969e+03      -1.532127e+04       1.004515e+09      -3.865120e+04      -7.528969e+03      -3.865120e+04      -7.753549e+08
