+++ tests/reference/berry_GaAs_W90-cumdos_iter-0000.dat
@@ -1,0 +1,12 @@
+#    Efermi                                               
+   7.000000e+00           7.787037e+00       7.794214e+00
+   7.200000e+00           7.879630e+00       7.880422e+00
+   7.400000e+00           7.981481e+00       7.974199e+00
+   7.600000e+00           7.981481e+00       7.982633e+00
+   7.800000e+00           8.000000e+00       7.998710e+00
+   8.000000e+00           8.000000e+00       8.000624e+00
+   8.200000e+00           8.009259e+00       8.008663e+00
+   8.400000e+00           8.009259e+00       8.012075e+00
+   8.600000e+00           8.046296e+00       8.054230e+00
+   8.800000e+00           8.194444e+00       8.189055e+00
+   9.000000e+00           8.268519e+00       8.262669e+00
