+++ tests/reference/berry_GaAs_W90-dos_iter-0000.dat
@@ -1,0 +1,12 @@
+#    Efermi                                               
+   7.000000e+00           4.629630e-01       4.643428e-01
+   7.200000e+00           4.861111e-01       4.678787e-01
+   7.400000e+00           2.546296e-01       2.562247e-01
+   7.600000e+00           4.629630e-02       6.127653e-02
+   7.800000e+00           4.629630e-02       4.497891e-02
+   8.000000e+00           2.314815e-02       2.488203e-02
+   8.200000e+00           2.314815e-02       2.862747e-02
+   8.400000e+00           9.259259e-02       1.139183e-01
+   8.600000e+00           4.629630e-01       4.429893e-01
+   8.800000e+00           5.555556e-01       5.340935e-01
+   9.000000e+00           3.472222e-01       3.628516e-01
